Share FactCollector pure helpers across Lite/Dashboard
Second wave of the code-sharing follow-up. The Dashboard SqlServerFactCollector
and Lite DuckDbFactCollector each carried byte-identical, data-source-agnostic
fact-shaping helpers. Moved them to a single PerformanceMonitor.Analysis class
(FactCollectorHelpers) so the two collectors can't drift:

- EmitServerHealthFacts (WS5 IFI/LPIM/memory-dump advisories) + the
  LpimAdvisoryMinPhysicalMemoryMb RAM-floor const
- GroupGeneralLockWaits / IsGeneralLockWait (LCK grouping)
- GroupParallelismWaits (CX* grouping)

These operate only on the already-collected Fact list + AnalysisContext, so no
new project dependencies are needed (the helper lives in Analysis, where Fact
and AnalysisContext already are). Both collectors now call FactCollectorHelpers.*.

Pure refactor, no behavior change. Solution builds clean; Dashboard.Tests (488),
Lite.Tests (544), Installer.Tests (80) all pass.
