+++ Lite/Analysis/DuckDbFactCollector.cs
@@ -28,8 +28,8 @@ public class DuckDbFactCollector : IFactCollector
         var facts = new List<Fact>();
 
         await CollectWaitStatsFactsAsync(context, facts);
-        GroupGeneralLockWaits(facts, context);
-        GroupParallelismWaits(facts, context);
+        FactCollectorHelpers.GroupGeneralLockWaits(facts, context);
+        FactCollectorHelpers.GroupParallelismWaits(facts, context);
         await CollectBlockingFactsAsync(context, facts);
         await CollectBlockingChainFactsAsync(context, facts);
         await CollectDeadlockFactsAsync(context, facts);
@@ -1891,73 +1891,9 @@ LIMIT 1";
             // WS5 server-health advisories (advise-only). Gating mirrors the Dashboard collector so
             // both apps agree on what is worth flagging; a fact that would score 0 is simply never
             // emitted (noise control).
-            EmitServerHealthFacts(context, facts, edition, physicalMemMb, lpim, ifi, dumpCount);
+            FactCollectorHelpers.EmitServerHealthFacts(context, facts, edition, physicalMemMb, lpim, ifi, dumpCount);
         }
         catch { /* Table may not exist or have no data */ }
-    }
-
-    // RAM floor below which LPIM-off is not worth flagging — shared rule with the Dashboard
-    // SqlServerFactCollector (small buffer pools do not suffer from OS paging the way large ones do).
-    private const long LpimAdvisoryMinPhysicalMemoryMb = 32 * 1024;
-
-    /// <summary>
-    /// Emits the WS5 advise-only server-health facts (IFI off / LPIM off / memory dumps) from the
-    /// latest server_properties values, applying the same noise-control gating as the Dashboard:
-    ///   • IFI: emit whenever the value is known (Value = enabled bit) — universally good advice.
-    ///   • LPIM: emit only on non-Express editions with meaningful RAM (Value = enabled bit).
-    ///   • Dumps: emit whenever the count is known (Value = count) — the scorer flags count > 0.
-    /// </summary>
-    private static void EmitServerHealthFacts(
-        AnalysisContext context, List<Fact> facts, string edition, long physicalMemMb,
-        bool? lockPagesInMemory, bool? instantFileInit, int? memoryDumpCount)
-    {
-        var isExpress = edition.Contains("Express", StringComparison.OrdinalIgnoreCase);
-
-        if (instantFileInit.HasValue)
-        {
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "CONFIG_IFI_DISABLED",
-                Value = instantFileInit.Value ? 1 : 0,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["instant_file_initialization_enabled"] = instantFileInit.Value ? 1 : 0
-                }
-            });
-        }
-
-        if (lockPagesInMemory.HasValue && !isExpress && physicalMemMb >= LpimAdvisoryMinPhysicalMemoryMb)
-        {
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "CONFIG_LPIM_DISABLED",
-                Value = lockPagesInMemory.Value ? 1 : 0,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["lock_pages_in_memory"] = lockPagesInMemory.Value ? 1 : 0,
-                    ["physical_memory_mb"] = physicalMemMb
-                }
-            });
-        }
-
-        if (memoryDumpCount.HasValue)
-        {
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "SERVER_MEMORY_DUMPS",
-                Value = memoryDumpCount.Value,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["memory_dump_count"] = memoryDumpCount.Value
-                }
-            });
-        }
     }
 
     /// <summary>
@@ -2104,117 +2040,6 @@ FROM latest WHERE rn = 1";
             });
         }
         catch { /* Table may not exist or have no data */ }
-    }
-
-    /// <summary>
-    /// Groups general lock waits (X, U, IX, SIX, BU, IU, UIX, etc.) into a single "LCK" fact.
-    /// Keeps individual facts for:
-    ///   - LCK_M_S, LCK_M_IS (reader/writer blocking — RCSI signal)
-    ///   - LCK_M_RS_*, LCK_M_RIn_*, LCK_M_RX_* (serializable/repeatable read signal)
-    ///   - SCH_M, SCH_S (schema locks — DDL/index operations)
-    /// Individual constituent wait times are preserved in metadata as "{type}_ms" keys.
-    /// </summary>
-    private static void GroupGeneralLockWaits(List<Fact> facts, AnalysisContext context)
-    {
-        var generalLocks = facts.Where(f => f.Source == "waits" && IsGeneralLockWait(f.Key)).ToList();
-        if (generalLocks.Count == 0) return;
-
-        var totalWaitTimeMs = generalLocks.Sum(f => f.Metadata.GetValueOrDefault("wait_time_ms"));
-        var totalWaitingTasks = generalLocks.Sum(f => f.Metadata.GetValueOrDefault("waiting_tasks_count"));
-        var totalSignalMs = generalLocks.Sum(f => f.Metadata.GetValueOrDefault("signal_wait_time_ms"));
-        var avgMsPerWait = totalWaitingTasks > 0 ? totalWaitTimeMs / totalWaitingTasks : 0;
-        var fractionOfPeriod = totalWaitTimeMs / context.PeriodDurationMs;
-
-        var metadata = new Dictionary<string, double>
-        {
-            ["wait_time_ms"] = totalWaitTimeMs,
-            ["waiting_tasks_count"] = totalWaitingTasks,
-            ["signal_wait_time_ms"] = totalSignalMs,
-            ["resource_wait_time_ms"] = totalWaitTimeMs - totalSignalMs,
-            ["avg_ms_per_wait"] = avgMsPerWait,
-            ["period_duration_ms"] = context.PeriodDurationMs,
-            ["lock_type_count"] = generalLocks.Count
-        };
-
-        // Preserve individual constituent wait times for detailed analysis
-        foreach (var lck in generalLocks)
-            metadata[$"{lck.Key}_ms"] = lck.Metadata.GetValueOrDefault("wait_time_ms");
-
-        // Remove individual facts, add grouped fact
-        foreach (var lck in generalLocks)
-            facts.Remove(lck);
-
-        facts.Add(new Fact
-        {
-            Source = "waits",
-            Key = "LCK",
-            Value = fractionOfPeriod,
-            ServerId = context.ServerId,
-            Metadata = metadata
-        });
-    }
-
-    /// <summary>
-    /// Groups all CX* parallelism waits (CXPACKET, CXCONSUMER, CXSYNC_PORT, CXSYNC_CONSUMER, etc.)
-    /// into a single "CXPACKET" fact. They all indicate the same thing: parallel queries are running.
-    /// Individual wait times are preserved in metadata for detailed analysis.
-    /// </summary>
-    private static void GroupParallelismWaits(List<Fact> facts, AnalysisContext context)
-    {
-        var cxWaits = facts.Where(f => f.Source == "waits" && f.Key.StartsWith("CX", StringComparison.Ordinal)).ToList();
-        if (cxWaits.Count <= 1) return;
-
-        var totalWaitTimeMs = cxWaits.Sum(f => f.Metadata.GetValueOrDefault("wait_time_ms"));
-        var totalWaitingTasks = cxWaits.Sum(f => f.Metadata.GetValueOrDefault("waiting_tasks_count"));
-        var totalSignalMs = cxWaits.Sum(f => f.Metadata.GetValueOrDefault("signal_wait_time_ms"));
-        var avgMsPerWait = totalWaitingTasks > 0 ? totalWaitTimeMs / totalWaitingTasks : 0;
-        var fractionOfPeriod = totalWaitTimeMs / context.PeriodDurationMs;
-
-        var metadata = new Dictionary<string, double>
-        {
-            ["wait_time_ms"] = totalWaitTimeMs,
-            ["waiting_tasks_count"] = totalWaitingTasks,
-            ["signal_wait_time_ms"] = totalSignalMs,
-            ["resource_wait_time_ms"] = totalWaitTimeMs - totalSignalMs,
-            ["avg_ms_per_wait"] = avgMsPerWait,
-            ["period_duration_ms"] = context.PeriodDurationMs
-        };
-
-        // Preserve individual constituent wait times for detailed analysis
-        foreach (var cx in cxWaits)
-            metadata[$"{cx.Key}_ms"] = cx.Metadata.GetValueOrDefault("wait_time_ms");
-
-        foreach (var cx in cxWaits)
-            facts.Remove(cx);
-
-        facts.Add(new Fact
-        {
-            Source = "waits",
-            Key = "CXPACKET",
-            Value = fractionOfPeriod,
-            ServerId = cxWaits[0].ServerId,
-            Metadata = metadata
-        });
-    }
-
-    /// <summary>
-    /// Returns true for general lock waits that should be grouped into "LCK".
-    /// Excludes reader locks (S, IS), range locks (RS_*, RIn_*, RX_*), and schema locks.
-    /// </summary>
-    private static bool IsGeneralLockWait(string waitType)
-    {
-        if (!waitType.StartsWith("LCK_M_", StringComparison.OrdinalIgnoreCase)) return false;
-
-        // Keep individual: reader/writer locks
-        if (waitType is "LCK_M_S" or "LCK_M_IS") return false;
-
-        // Keep individual: range locks (serializable/repeatable read)
-        if (waitType.StartsWith("LCK_M_RS_", StringComparison.OrdinalIgnoreCase) ||
-            waitType.StartsWith("LCK_M_RIn_", StringComparison.OrdinalIgnoreCase) ||
-            waitType.StartsWith("LCK_M_RX_", StringComparison.OrdinalIgnoreCase)) return false;
-
-        // Everything else (X, U, IX, SIX, BU, IU, UIX, etc.) → group
-        return true;
     }
 
     private static long ToInt64(object value)
